изменит названия файлов и удалит ненужные
