+++ js/main.js
@@ -1,23 +1,5 @@
-import {createDescriptions} from './descriptions.js';
+import {createDescriptions} from './utils.js';
+import {messages, names, photoDescription} from './data.js';
 
-const messages = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const names = [ 'Иван', 'Илья', 'Сергей', 'Георгий', 'Максим', 'Виктория', 'Алёна', 'Ольга', 'Наталья', 'Яна' ];
-
-const photoDescription = [
-  'Неспокойное море',
-  'Смешной котёнок',
-  'Цветочный сад',
-  'Ночной мегаполис',
-  'Кафе на берегу',
-  'Цветочный сад'
-];
 
 createDescriptions(photoDescription, messages, names);
